Rename to UpdateTagStep
